requirements.txt
black
hydra-ax-sweeper
hydra-joblib-launcher
hydra-core
hydra_colorlog
matplotlib
numpy
torch
tqdm
wandb